Ruttenberg/helium packet router issue 004 (#25)
* interim: gwmp client

* interim: add packet record

* interim: push_data and some callbacks.

* Pass Handler Pid to push_data function.

* GWMP worker handles UDP responses.

* Supervisor for GWMP UDP worker.

* Move gwmp/ into protocols/.

* Take router_utils modules.

* Change interface of hpr_gwmp_client:push_data/4.

* Join working for virtual device

- add some deps [iso8601, xxhash]
- turn packet_up to rxpk
- turn txpk to packet_down

Make sure hotspots start with a valid address to get PULL_DATA working.
Even if you send a PUSH_DATA, until an LNS receives a PULL_DATA it will
not think it can send a PULL_RESP.

More work on the PULL_RESP front needs to be done. Make sure we
inititititiate (Billy Joel) PULL_RESP to unseen LNS's so they can
respond to this specific gateway.

* Fix bugs. Move various utils functions.

* Correct the worker module name.

* client -> worker

* Correct type for Handlers will always be a stream

* rename socket_info() -> socket_dest()

implies more the data structures contents

* add path for routing gwmp protocol

Move the gwmp routing to its own file, and send all grpc to hpr_routing
for dispatching.

NOTE: Streams should probably be handled in hpr_routing.

* Updates to gwmp worker

- Handle PULL_DATA to multiple LNS at the same time.
- general clean up.
- drop metrics handlers for future.
- rename #state.

* formatting

* fix semtech_udp tests, when did return_maps stop being the default?

* inline most of the udp sup

* fix error type from udp_sup

* fully remove udp metrics

We'll have a good hard think about what's important for us to know with
udp workers, and get those metrics setup.

* Remove #state.location

* Call hpr_gwmp_send_response directly.

* Fix type error for StreamHandler.

* rename HandlerPid -> StreamHandler

* log errors, but don't send error message

We have a streamhandler here, and we want to close the streams if
something goes wrong (yet?). For now, we log and continue.

* spawn handle_packet to move as fast as possible

"gotta go fast" - STH

* use spec from behavior

* update headers with company info

* no longer sending messages to reponse pid

don't expect it in the test

* return errors to the packet deliverer

* remove unused

* inline pull_resp handling

* return constructed tuples

* use the records

* downlink test

contains rudimentary grpc send mock

* factor out test packet up

* test ack Downlink from multiple LNS

* replace pid with atom to blow all the way up if its used

also a typo

* remove gwmp_udp_socket

it was a thin wrapper around gen_udp, and didn't contain any useful
features. Also allowed us to remove the prefilling of a socket with a
destination.

* typo; format

* Use explicit refs/tags for dependencies.

* use proc_lib spawn

this will give us a crash report is something bad happens.

* removed unused utils

* remove noisy docs from template

* don't log entire payload

* make sure gws start up uniquely

* update spec for service

had an old version of the plugin that generated the wrong spec

Co-authored-by: Michael Jeffrey <michaeldjeffrey@gmail.com>