Remove all routes when a stream is successfully obtained
Packet Router starts up by trying to load routes from a backup file. If
a Route in that file was deleted while there was no connection to the
config service, packet router would never no to get rid of it.

When a new connection is made, config service will send _all_ routes, we
can no throw away the backups and start fresh.