monitor pmap spawns (#204)
If a pmap thread goes down, the process that asked for the pmap will
hang forever.

Here, we monitor all the spawned processes, and count them against the
expected number of results.

The down message only applies to a single process.
Worst case scenario, on a box with few cores, and many routes to
deliver, one route can go down, and kill the delivery for following
routes. Or prevent reporting for routes that were already delivered.

This PR is addressing a process leak.