routing by packet lookup (#257)
* move multi buy key to multi buy module

* further constrict multi buy request test

do not check how many requests have been sent until the multi-buy
function has been called enough times that it could be more than what we expect.

* queue packets in ETS until routing is determined

When it packet comes into routing...
1. It's never been seen before and we need to figure out the routing
   information for it.
2. We've seen it before, and have not yet determined the routing, in
   which case it get's queued.
3. We've seen it before, and already know what to do.

* break out make skf ets tables

* stop using timestamp in skf check

* No longer treating SKF as ordered set

SKF tables are not sets with no timing information. We can do direct
inserts and deletes for greater performance.

* add packet-routing cleanup job, return all applicable skf routes

SKF are unique in their route tables since the last_used key has been
removed. If more than one route has an applicable skf, let's send to it.

* add routing_cleanup to sys.config

every 15 seconds, anything older than 2 minutes will be cleaned out.

* reset matchspec field numbers

* attempt at simplifying routin cleanup

Do not let more than 1 cleanup routine start.
Check if the current cleanup is registered, and bail if there's another
PID claiming to be the cleanup.

* better handle race conditions on acquiring a lock

when more than one packet is told to establish routing, the first will
get the lock and continue. The second will see it's not the first entry,
and go back to looking up routing again. Most likely to queue itself.

* dust the shelves, remove old comment

* move routing cache to it's own worker

* add a wait branch that gives more information about the failure

* update cleanup test for new table name

* use tuple for dialyzer

* remove stream status

inserts and deletes are from a set table with a known key, not an
ordered_set with a timestamp as the key.