rehydrate since (#276)
* move cli functions to their modules

* config rehydrate, and updates since

- rehydrate config from dets files upon startup
- when starting the config update stream, send the timestamp of the last
  update we recieved

- todo: if any of the config cannot be rehydrated, remove everything and
start from the beginning.

* fix table names and tests passing

use with_open_dets helpers. Keeping them in individual files so it
doesn't become unneccassarily generic

* add some timing around ets initialization in supervisor

* do not spawn skf rehydration

to avoid conflicts of rehydrating while updates are streaming in.

* make sure all the ets/dets config is consistent

* remove from merge

* rebase test service rename

* process routes before other updates

* Start the checkpoint timer when a stream is successfully initialized.

maybe cancel any previouss timers so we don't build up checkpoints if a
disconnect happens.

* don't get caught in a checkpoint loop or race condition

- add helpers for manually running checkpoints.
- cancel timers correctly.
- Don't make a timer until we have a connection.

* helpers for checking next checkpoint

* delete dets file when route is deleted

* Route Delete, ensure skf file deletion

A Route Delete can come through as part of the restreaming of events.
The Route may not exist in the ets table, and we want to make sure teh
skf file does not stick around.

* HPR only deals in milliseconds

timestamps from routing responses are unix epoch seconds, requests are
sent in milliseconds.

* helper counts function for metrics CLI

* fix counts table name

* return message to log and print

* reset connection by killing the worker

* add cli commands

* proto has been merged upstream