Fix parsing http downlink tokens
It would occur sometimes that a pubkeybin would contain the separator,
leading to a malformed_token because there were too many parts of the
token. Or the region would start with part of the separator and things
would break when attempting to turn it into an atom.

Putting the PubKeyBin in the last position allows us to split the token
until we know we have all the parts, and be sure the rest of the binary
is the PubKeyBin, not caring if it contains the seperator.