move syncing routes to cli module
- add cli config test suite
- pretty print the added and removed ids or routes

The diffing for the routes is done in the cli module, this way the route
stream worker doesn't start to become like the xor filter worker where
it started to hold the kitchen sink.

Routes are only sent to the route stream worker after they have been
inserted and need their parts updated.